ui tests for login, register, upload feed
