+++ app/src/androidTest/java/com/htetarkarzaw/twitterlite/ui_test/TwitterLiteUiTest.kt
@@ -1,0 +1,125 @@
+package com.htetarkarzaw.twitterlite.ui_test
+
+import androidx.navigation.findNavController
+import androidx.test.core.app.ActivityScenario
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.action.ViewActions.typeText
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.contrib.RecyclerViewActions
+import androidx.test.espresso.matcher.ViewMatchers.isDescendantOfA
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import com.htetarkarzaw.twitterlite.R
+import com.htetarkarzaw.twitterlite.ui.MainActivity
+import com.htetarkarzaw.twitterlite.ui.screen.feed.FeedAdapter
+import org.hamcrest.CoreMatchers.allOf
+import org.junit.Assert.*
+import org.junit.Test
+import org.junit.runner.RunWith
+import kotlin.random.Random
+import kotlin.random.nextInt
+
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class TwitterLiteUiTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("com.htetarkarzaw.twitterlite", appContext.packageName)
+    }
+
+    private val activityScenario: ActivityScenario<MainActivity>? = ActivityScenario.launch(MainActivity::class.java)
+
+    @Test
+    fun registerFlow() {
+        Thread.sleep(3000L)
+        var currentDestinationId: Int? = 0
+        activityScenario?.onActivity { activity ->
+            val navController = activity.findNavController(R.id.nav_host_fragment_activity_main)
+            currentDestinationId = navController.currentDestination?.id
+        }
+        if (currentDestinationId == R.id.feedFragment) {
+            onView(allOf(withId(R.id.settingFragment), isDescendantOfA(withId(R.id.navView)))).perform(click())
+            onView(withId(R.id.cvLogout)).perform(click())
+            Thread.sleep(3000L)
+        }
+        onView(withId(R.id.btnRegister)).perform(click())
+        val random = Random.nextInt(0..100)
+        onView(withId(R.id.etEmail)).perform(typeText("arkar$random@gmail.com"))
+        onView(withId(R.id.etDisplayName)).perform(typeText("Htet Arkar Zaw $random"))
+        onView(withId(R.id.etPassword)).perform(typeText("12345678"))
+        onView(withId(R.id.etConfirmPassword)).perform(typeText("12345678"))
+        onView(withId(R.id.btnRegister)).perform(click())
+        Thread.sleep(5000L)
+        activityScenario?.onActivity { activity ->
+            val navController = activity.findNavController(R.id.nav_host_fragment_activity_main)
+            currentDestinationId = navController.currentDestination?.id
+
+            // Compare the current destination ID with the expected destination ID
+            assertEquals(R.id.feedFragment, currentDestinationId)
+        }
+    }
+    
+    @Test
+    fun loginFlow() {
+        Thread.sleep(3000L)
+        var currentDestinationId: Int? = 0
+        activityScenario?.onActivity { activity ->
+            val navController = activity.findNavController(R.id.nav_host_fragment_activity_main)
+            currentDestinationId = navController.currentDestination?.id
+        }
+        if (currentDestinationId == R.id.feedFragment) {
+            onView(allOf(withId(R.id.settingFragment), isDescendantOfA(withId(R.id.navView)))).perform(click())
+            onView(withId(R.id.cvLogout)).perform(click())
+            Thread.sleep(3000L)
+        }
+        onView(withId(R.id.etEmail)).perform(typeText("arkarzaw412@gmail.com"))
+        onView(withId(R.id.etPassword)).perform(typeText("12345678"))
+        onView(withId(R.id.btnLogin)).perform(click())
+        Thread.sleep(5000L)
+        activityScenario?.onActivity { activity ->
+            val navController = activity.findNavController(R.id.nav_host_fragment_activity_main)
+            currentDestinationId = navController.currentDestination?.id
+            assertEquals(R.id.feedFragment, currentDestinationId)
+        }
+    }
+
+
+    @Test
+    fun uploadFeedFlow() {
+        var currentDestinationId: Int? = 0
+        Thread.sleep(3000L)
+        onView(withId(R.id.fab)).perform(click())
+        onView(withId(R.id.etTweet)).perform(typeText("This is Post From Instrumented Testing"))
+        onView(withId(R.id.btnUpload)).perform(click())
+        Thread.sleep(5000L)
+        activityScenario?.onActivity { activity ->
+            val navController = activity.findNavController(R.id.nav_host_fragment_activity_main)
+            currentDestinationId = navController.currentDestination?.id
+
+            assertEquals(
+                R.id.feedFragment,
+                currentDestinationId
+            )
+        }
+        if (currentDestinationId == R.id.feedFragment) {
+            onView(withId(R.id.rvFeed)).perform(
+                RecyclerViewActions.actionOnItemAtPosition<FeedAdapter.FeedViewHolder>(
+                    0,
+                    click()
+                )
+            )
+            onView(withId(R.id.tvTweet)).check(matches(withText("This is Post From Instrumented Testing")))
+        }
+    }
+
+}
